Split accounting close cockpit view model tests
